Refactor: Move AlignmentOptimizer to judges/base.py to eliminate circular imports
This refactoring improves the module structure by consolidating related classes:

- Move AlignmentOptimizer class from separate file to mlflow/genai/judges/base.py
  - Co-locates with Judge class to eliminate circular import risks
  - Maintains all existing functionality and API
  - Same @experimental decorator and method signature

- Remove mlflow/genai/judges/alignment_optimizer.py (no longer needed)
  - Consolidated into base.py for cleaner structure
  - Eliminates potential circular import issues

- Update mlflow/genai/judges/__init__.py imports
  - Import both AlignmentOptimizer and Judge from .base module
  - Simplified import structure - single source for related classes
  - Public API remains unchanged

- Verify all functionality preserved
  - All tests continue to pass (5/5 passing)
  - No breaking changes to existing Judge functionality
  - Import paths remain the same for end users

Benefits:
- Eliminates circular import risks
- Cleaner module organization
- Simplified maintenance
- Better cohesion of related classes
